Descargar datos meteorológicos

La función descargar_datos toma el ID de una estación y un directorio de destino, y descarga en dicha ubicación un archivo CSV con los datos meteorológicos de la estación correspondiente. Los IDs posibles son: NH0098, NH0046, NH437, NH472 y NH0910.

Usage

descargar_datos(id_estacion, directorio_destino)

Arguments

id_estacion

ID de la estación.

directorio_destino

Directorio donde se guardará el archivo CSV.

Value

Un archivo CSV con los datos meteorológicos de la estación con el ID ingresado.

Examples

descargar_datos("NH0098", "data")
#> Warning: One or more parsing issues, call `problems()` on your data frame for details,
#> e.g.:
#>   dat <- vroom(...)
#>   problems(dat)
#> Rows: 23733 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (2): id, direccion_viento_1000cm
#> dbl  (21): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (11): temperatura_intemperie_50cm_minima, temperatura_inte_5cm, tempera...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 23,733 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0098 1956-04-01                     13.6                            21.6
#>  2 NH0098 1956-04-02                     16.8                            23.6
#>  3 NH0098 1956-04-03                     19                              25.9
#>  4 NH0098 1956-04-04                     17.6                            23.6
#>  5 NH0098 1956-04-05                     20                              27.2
#>  6 NH0098 1956-04-06                     19.2                            24.7
#>  7 NH0098 1956-04-07                     15                              19.1
#>  8 NH0098 1956-04-08                     18.8                            26.8
#>  9 NH0098 1956-04-09                     18.2                            26.6
#> 10 NH0098 1956-04-10                     11.4                            15.4
#> # ℹ 23,723 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <dbl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <dbl>, temperatura_suelo_10cm_media <dbl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <dbl>, …
descargar_datos("NH0910", "data")
#> Rows: 15553 Columns: 35
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr   (1): id
#> dbl   (5): temperatura_abrigo_150cm, temperatura_abrigo_150cm_maxima, temper...
#> lgl  (28): temperatura_intemperie_5cm_minima, temperatura_intemperie_50cm_mi...
#> date  (1): fecha
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 15,553 × 35
#>    id     fecha      temperatura_abrigo_150cm temperatura_abrigo_150cm_maxima
#>    <chr>  <date>                        <dbl>                           <dbl>
#>  1 NH0910 1975-01-01                     21                              31  
#>  2 NH0910 1975-01-02                     21.2                            28.5
#>  3 NH0910 1975-01-03                     24.2                            32.5
#>  4 NH0910 1975-01-04                     23.5                            31  
#>  5 NH0910 1975-01-05                     19                              22  
#>  6 NH0910 1975-01-06                     22.2                            27.5
#>  7 NH0910 1975-01-07                     24.2                            30  
#>  8 NH0910 1975-01-08                     26.5                            33.5
#>  9 NH0910 1975-01-09                     26.5                            35  
#> 10 NH0910 1975-01-10                     19.2                            25  
#> # ℹ 15,543 more rows
#> # ℹ 31 more variables: temperatura_abrigo_150cm_minima <dbl>,
#> #   temperatura_intemperie_5cm_minima <lgl>,
#> #   temperatura_intemperie_50cm_minima <lgl>,
#> #   temperatura_suelo_5cm_media <lgl>, temperatura_suelo_10cm_media <lgl>,
#> #   temperatura_inte_5cm <lgl>, temperatura_intemperie_150cm_minima <lgl>,
#> #   humedad_suelo <lgl>, precipitacion_pluviometrica <dbl>, granizo <lgl>, …