Crea una Tabla Resumen de la Temperatura Abrigo

La función tabla_resumen_temperatura crea una tabla de resumen de la temperatura_abrigo_150cm para una o más estaciones.

Usage

tabla_resumen_temperatura(datos)

Arguments

datos

data.frame de las estaciones de cuya temperatura abrigo se quiere crear una tabla resumen.

Value

Devuelve una tabla resumen de la temperatura abrigo de las estaciones ingresadas.

Examples

# Cargar los datos de ejemplo desde el paquete
data("estaciones_merged")

# Llamar a la función usando los datos cargados
tabla_resumen_temperatura(estaciones_merged)
#> # A tibble: 5 × 4
#>   id     min_temp max_temp mean_temp
#>   <chr>     <dbl>    <dbl>     <dbl>
#> 1 NH0046    -0.25     34.6      17.3
#> 2 NH0098   -17.4      37.4      18.6
#> 3 NH0437     2.1      36.0      20.1
#> 4 NH0472     0.55     34.8      18.0
#> 5 NH0910     0        35.2      17.7